refactor : bookCard, bookGuide 폴더 생성
#636
